requirements.txt
# Web笔记应用原生部署依赖包
# 基于AutoGen 0.7.1框架的生产环境配置

# ========================================
# AutoGen核心框架 (0.7.1版本)
# ========================================
autogen-agentchat==0.7.1
autogen-core==0.7.1
autogen-ext==0.7.1

# ========================================
# Web服务框架
# ========================================
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
pydantic>=2.10,<3.0
starlette==0.27.0

# ========================================
# AI模型和API客户端
# ========================================
openai==1.3.7
dashscope==1.17.0
anthropic==0.7.8

# ========================================
# 向量数据库和搜索
# ========================================
chromadb==0.4.18
sentence-transformers==2.2.2
faiss-cpu==1.7.4

# ========================================
# 数据库和存储
# ========================================
sqlalchemy==2.0.23
alembic==1.12.1
# sqlite3 为Python标准库，勿通过pip安装

# ========================================
# 文档处理和解析
# ========================================
python-multipart==0.0.6
markdown==3.5.1
beautifulsoup4==4.12.2
pypdf2==3.0.1

# ========================================
# 搜索工具集成
# ========================================
google-api-python-client==2.108.0
requests==2.31.0

# ========================================
# 配置和环境管理
# ========================================
python-dotenv==1.0.0
pyyaml==6.0.1
toml==0.10.2

# ========================================
# 日志和监控
# ========================================
loguru==0.7.2
prometheus-client==0.19.0

# ========================================
# 系统工具和实用程序
# ========================================
psutil==5.9.6
click==8.1.7
rich==13.7.0
typer==0.9.0

# ========================================
# 安全和认证
# ========================================
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# ========================================
# 异步和并发
# ========================================
# asyncio 为Python标准库，勿通过pip安装
aiofiles==23.2.1
httpx>=0.28.1,<1
